Add GitHub Action and test workflow
Add GitHub Action which can be invoked from this or
downstream repositories to build the package and
its artifacts for subsequent use.

Add a GitHub Workflow to test that this build process
is successful, that will be triggered whenever code
or config changes are pushed to the remote repository
or when pull requests are submitted.