Remove unit test that requires network connectivity
Remove the default constructor unit test for
ConsentStreamIngestor, which can only be run on hosts
that have network connectivity + have AWS credentials
and permissions to query the expected DynamoDB table.

It's sufficient for our use cases to limit unit tests
to those that mock the DynamoDB client so no network
connectivity is required during unit tests.